requirements.txt
discord.py
pymongo
dsnpython
python-dotenv
